Refactored the implementation of parameterized tests macros to internal/
